Created subclasses of Being
