+++ LittlePrince.java
@@ -6,24 +6,14 @@ import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
  * @Zhiyu (Jennifer) Zhou
  * @Jan 7, 2022 - v1.0
  */
-public class LittlePrince extends Being
+public class LittlePrince extends Moving
 {
-    private int speed;
-    private Planet touchingPlanet;
     /**
      * Act - do whatever the LittlePrince wants to do. This method is called whenever
      * the 'Act' or 'Run' button gets pressed in the environment.
      */
     public void act()
     {
-        if (getOneIntersectingObject(Planet.class) != null){
-            touchingPlanet = (Planet)getOneIntersectingObject(Planet.class);
-            rotate(touchingPlanet.getRadius());
-        }
-    }
-
-    public void rotate(double planet){
-        move(planet);
-        turn(0.5);
+        super.act();
     }
 }
