+++ Moving.ctxt
@@ -1,0 +1,9 @@
+#BlueJ class context
+comment0.target=Moving
+comment0.text=\n\ Write\ a\ description\ of\ class\ Moving\ here.\n\ \n\ @author\ (your\ name)\ \n\ @version\ (a\ version\ number\ or\ a\ date)\n
+comment1.params=
+comment1.target=void\ act()
+comment1.text=\n\ Act\ -\ do\ whatever\ the\ Character\ wants\ to\ do.\ This\ method\ is\ called\ whenever\n\ the\ 'Act'\ or\ 'Run'\ button\ gets\ pressed\ in\ the\ environment.\n
+comment2.params=planet
+comment2.target=void\ rotate(double)
+numComments=3
