+++ Stationary.ctxt
@@ -1,0 +1,7 @@
+#BlueJ class context
+comment0.target=Stationary
+comment0.text=\n\ Write\ a\ description\ of\ class\ Stationary\ here.\n\ \n\ @author\ (your\ name)\ \n\ @version\ (a\ version\ number\ or\ a\ date)\n
+comment1.params=
+comment1.target=void\ act()
+comment1.text=\n\ Act\ -\ do\ whatever\ the\ Stationary\ wants\ to\ do.\ This\ method\ is\ called\ whenever\n\ the\ 'Act'\ or\ 'Run'\ button\ gets\ pressed\ in\ the\ environment.\n
+numComments=2
