+++ Stationary.java
@@ -1,18 +1,19 @@
 import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
 
 /**
- * Write a description of class Character here.
+ * Write a description of class Stationary here.
  * 
  * @author (your name) 
  * @version (a version number or a date)
  */
-public class Being extends SuperSmoothMover
+public class Stationary extends Being
 {
     /**
-     * Act - do whatever the Character wants to do. This method is called whenever
+     * Act - do whatever the Stationary wants to do. This method is called whenever
      * the 'Act' or 'Run' button gets pressed in the environment.
      */
     public void act()
     {
+        // Add your action code here.
     }
 }
